fixed devpi-server --stop on Windows
This allows devpi-server --stop to work on Windows, which it did not
do before. By doing that, however, it does not work anymore with GUI
applications, if it ever did before.

Windows has no equivalent of fork(). Therefore, setuptools uses a
combination of a launcher binary and a wrapper script in place of
console scripts. Every time a console "script" is invoked, the
launcher binary starts a new python process, which executes the
launcher script, which does whatever the console script would have
done in the first place.

This approach has the downside that when devpi-server --start invokes
itself to detach from the console, the PID it gets back is that of the
launcher process. Killing that has no effect on the actual server
process (a child process of the launcher).

The launcher binary is provided by setuptools, and offers no means at
all to tell it to pass back the PID of its child process securely,
while remaining compatible with previous releases.

This change does two things:

- It creates the child process in a new process group on Windows, as
  already happens on *nix. This prevents Ctrl-C keypresses in the
  console window from reaching the child process.

- It uses os.kill(pid, signal.CTRL_BREAK_EVENT) to kill the child
  process. Starting with Python 2.7 and 3.2, this is supported on
  Windows.

  CTRL_BREAK_EVENT is different from CTRL_C_EVENT; the latter cannot
  be sent to another process group. While the setuptools launcher
  binary will forward CTRL_C_EVENT to its child process, it ignores
  CTRL_BREAK_EVENT. This is not a problem, because it will be received
  by all processes in the process group, and they will exit
  independently of each other.

When run in the foreground (without --start), there is never a change
in process groups, and pressing Ctrl-C works in the normal way.

The previous approach (which simply terminates the target process) is
left in place, just in case. It will never be used on Python 2.7+ and
3.2+.

All tests pass on:

- Python 2.7.6 (default, Nov 10 2013, 19:24:24)
  [MSC v.1500 64 bit (AMD64)] on win32
- Python 3.3.3 (v3.3.3:c3896275c0f6, Nov 18 2013, 21:19:30)
  [MSC v.1600 64 bit (AMD64)] on win32